feat: prevent field deletion if used in templated string fields
