+++ designer/src/components/condition.tsx
@@ -100,37 +100,6 @@ export function isFieldUsedInCondition(
   return false;
 }
 
-/**
- * Given the entire fields/fviews object, returns all places
- * (sections or fields) that reference `fieldName` in a condition.
- */
-export function findFieldUsage(
-  fieldName: string,
-  allFields: Record<string, any>,
-  allFviews: Record<string, any>
-): string[] {
-  const affected: string[] = [];
-
-  // Check section-level conditions
-  for (const sectionId in allFviews) {
-    const condition = allFviews[sectionId].condition;
-    if (isFieldUsedInCondition(condition, fieldName)) {
-      affected.push(`Section: ${allFviews[sectionId].label}`);
-    }
-  }
-
-  // Check field-level conditions
-  for (const fId in allFields) {
-    const condition = allFields[fId].condition;
-    if (isFieldUsedInCondition(condition, fieldName)) {
-      const label = allFields[fId]['component-parameters']?.label ?? fId;
-      affected.push(`Field: ${label}`);
-    }
-  }
-
-  return affected;
-}
-
 export const ConditionModal = (props: ConditionProps & {label: string}) => {
   const [open, setOpen] = useState(false);
