+++ designer/src/components/field-editor.tsx
@@ -52,7 +52,7 @@ import {TakePhotoFieldEditor} from './Fields/TakePhotoField';
 import {TemplatedStringFieldEditor} from './Fields/TemplatedStringFieldEditor';
 import {TextFieldEditor} from './Fields/TextFieldEditor';
 import {useState, useMemo} from 'react';
-import {findFieldUsage} from './condition';
+import {isFieldUsedInCondition} from './condition';
 
 type FieldEditorProps = {
   fieldName: string;
@@ -100,11 +100,9 @@ export const FieldEditor = ({
   const [deleteWarningOpen, setDeleteWarningOpen] = useState(false);
   const [conditionsAffected, setConditionsAffected] = useState<string[]>([]);
 
-  // Called by a button onClick
   const deleteField = (evt: React.SyntheticEvent) => {
     evt.stopPropagation();
 
-    // ❌ NO HOOK calls here, but we can do normal function calls:
     const usage = findFieldUsage(fieldName, allFields, allFviews);
 
     if (usage.length > 0) {
@@ -178,6 +176,48 @@ export const FieldEditor = ({
       moveFieldCallback(targetViewId);
       handleCloseMoveDialog();
     }
+  };
+
+  /**
+   * Finds where a field is used in conditions or templated string fields
+   */
+  const findFieldUsage = (
+    fieldName: string,
+    allFields: Record<string, any>,
+    allFviews: Record<string, any>
+  ): string[] => {
+    const affected: string[] = [];
+
+    // ✅ Check section-level conditions
+    for (const sectionId in allFviews) {
+      const condition = allFviews[sectionId].condition;
+      if (isFieldUsedInCondition(condition, fieldName)) {
+        affected.push(`Section: ${allFviews[sectionId].label}`);
+      }
+    }
+
+    // ✅ Check field-level conditions
+    for (const fId in allFields) {
+      const condition = allFields[fId].condition;
+      if (isFieldUsedInCondition(condition, fieldName)) {
+        const label = allFields[fId]['component-parameters']?.label ?? fId;
+        affected.push(`Field Condition: ${label}`);
+      }
+    }
+
+    // ✅ Check for Templated String Fields using the deleted field
+    for (const fId in allFields) {
+      if (allFields[fId]['component-name'] === 'TemplatedStringField') {
+        const template = allFields[fId]['component-parameters']?.template || '';
+
+        if (template.includes(`{{${fieldName}}}`)) {
+          const label = allFields[fId]['component-parameters']?.label ?? fId;
+          affected.push(`Templated String: ${label} (uses '{{${fieldName}}}')`);
+        }
+      }
+    }
+
+    return affected;
   };
 
   // memoize the form value
